cyberrisk/NR/Ildiko/requirements.txt
absl-py==1.4.0
asttokens==2.2.1
astunparse==1.6.3
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.11.1
blis==0.7.9
cachetools==5.3.0
catalogue==2.0.8
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confection==0.0.4
constantly==15.1.0
contourpy==1.0.7
cryptography==43.0.1
cssselect==1.2.0
cupy-cuda11x==12.1.0
cycler==0.11.0
cymem==2.0.7
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
executing==1.2.0
fastrlock==0.8.1
fasttext==0.9.2
filelock==3.9.0
flatbuffers==23.3.3
fonttools==4.43.0
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.53.2
h5py==3.8.0
huggingface-hub==0.11.1
hyperlink==21.0.0
icecream==2.1.3
idna==3.7
incremental==22.10.0
isodate==0.6.1
itemadapter==0.7.0
itemloaders==1.0.6
jax==0.4.8
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.2.0
keras==2.13.1
keras-contrib @ git+https://www.github.com/keras-team/keras-contrib.git@3fc5ef709e061416f4bc8a92ca3750c824b5d2b0
keybert==0.7.0
kiwisolver==1.4.4
langcodes==3.3.0
Levenshtein==0.20.9
libclang==16.0.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
ml-dtypes==0.1.0
murmurhash==1.0.9
mwparserfromhell==0.6.4
networkx==3.0
nltk==3.9
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
parsel==1.7.0
pathy==0.10.1
Pillow==10.3.0
preshed==3.0.8
Protego==0.2.1
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.10.4
pycparser==2.21
pydantic==1.10.13
PyDispatcher==2.0.6
Pygments==2.15.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pytextrank==3.2.4
python-dateutil==2.8.2
python-version==0.0.2
pytz==2022.7.1
pywikibot==8.1.2
PyYAML==6.0
queuelib==1.6.2
rapidfuzz==2.13.7
rdflib==6.3.2
regex==2022.10.31
requests==2.32.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rich==13.2.0
rsa==4.9
scikit-learn==1.5.0
scipy==1.10.1
Scrapy==2.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.97
service-identity==21.1.0
six==1.16.0
sklearn==0.0.post4
smart-open==6.3.0
soupsieve==2.3.2.post1
spacy==3.5.3
spacy-alignments==0.9.0
spacy-fastlang==1.0.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spacy-transformers==1.2.4
SPARQLWrapper==2.0.0
srsly==2.4.5
tensorboard==2.12.1
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.2.0
thinc==8.1.10
threadpoolctl==3.1.0
tldextract==3.4.0
tokenizers==0.13.2
torch==2.2.0
torchvision==0.14.1
tqdm==4.66.3
transformers==4.38.0
Twisted==24.7.0
typer==0.7.0
typing_extensions==4.4.0
urllib3==1.26.19
w3lib==2.1.1
wasabi==1.1.1
Werkzeug==3.0.3
wrapt==1.14.1
zope.interface==5.5.2
